requirements.txt
joblib
matplotlib
numpy
pandas
Pillow
scikit-image
streamlit
streamlit-drawable-canvas
tensorflow
xgboost